678: Add contracted hours to inserts and tests
Also, update float data types to decimal

+++ migrations/app/20170508150532_app_workload_owner.js
@@ -3,7 +3,7 @@ exports.up = function (knex, Promise) {
   return knex.schema.createTable('workload_owner', function (table) {
     table.increments('id')
     table.integer('offender_manager_id').unsigned().notNullable().references('offender_manager.id')
-    table.float('contracted_hours').unsigned()
+    table.decimal('contracted_hours').unsigned()
     table.integer('team_id').unsigned().notNullable().references('team.id')
   }).catch(function (error) {
     console.log(error)
