+++ migrations/app/20170508150724_app_workload_points_calculations.js
@@ -11,8 +11,8 @@ exports.up = function (knex, Promise) {
     table.integer('paroms_points').unsigned().notNullable()
     table.integer('nominal_target').unsigned().notNullable()
     table.integer('available_points').unsigned().notNullable()
-    table.float('reduction_hours').defaultTo(0).notNullable()
-    table.float('contracted_hours').notNullable()
+    table.decimal('reduction_hours').defaultTo(0).notNullable()
+    table.decimal('contracted_hours').notNullable()
     table.timestamp('effective_from')
     table.timestamp('effective_to')
   }).catch(function (error) {
